Implemented the most important authorization rules

+++ pages/api/task/[id].js
@@ -1,7 +1,21 @@
 const models = require("@lib/server");
 const { Task } = models.default;
+import config from "@lib/sessionConfig";
+import { withIronSessionApiRoute } from "iron-session/next";
 
-export default async function handler(req, res) {
+export default withIronSessionApiRoute(handler, config);
+
+const handleNotAdminOrLeader = async (req, res) => {
+    if (!req.session.user.roles.find(role => {
+        return role.role === "admin" || role.role === "leader";
+    })) {
+        res.status(401).json({ message: "Unauthorized" });
+        return true;
+    }
+    return false;
+};
+
+export async function handler(req, res) {
     const {
         query: { id },
         method
@@ -11,9 +25,11 @@ export default async function handler(req, res) {
         res.status(200).json(await Task.findByPk(id));
         break;
     case "DELETE":
+        if (handleNotAdminOrLeader()) return;
         res.status(200).json(await Task.destroy({ where: { id }}));
         break;
     case "PUT":
+        if (handleNotAdminOrLeader()) return;
         res.status(200).json(await Task.update(req.body, { where: { id } }));
         break;
     default:
